+++ common.h
@@ -1,11 +1,12 @@
 // Needed for FILE*
 #include <stdio.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-extern size_t bytecount;
+extern uintmax_t bytecount;
 extern int acls, verbose;
 int create(FILE *dumpfile, FILE *tarfile);
 int extract(FILE *tarfile, FILE *dumpfile);
